feat: switch to tsd (#22)
* feat: switch to tsd

* feat: missing assertions

* fix: review

* Update types/types.test-d.ts
