+++ package.json
@@ -5,7 +5,7 @@
   "main": "lib/fastifySession.js",
   "scripts": {
     "unit": "ava test/*.test.js",
-    "tsc-check": "tsc --project ./types/tsconfig.json --noEmit",
+    "tsc-check": "tsd",
     "test": "npm run lint && npm run unit && npm run tsc-check",
     "lint": "standard lib/* test/*",
     "coveralls": "nyc npm test",
@@ -37,6 +37,7 @@
     "got": "^11.6.0",
     "nyc": "^15.0.0",
     "standard": "^16.0.1",
+    "tsd": "^0.17.0",
     "typescript": "^4.0.2"
   },
   "types": "types/types.d.ts",
